Remove GET_SIGNING_CERTIFICATES/GET_SIGNATURES from getPackageInfo
This change removes the `GET_SIGNING_CERTIFICATES` and `GET_SIGNATURES` flags from `PackageManager.getPackageInfo` calls in `PackageService.kt`. These flags are resource-intensive as they trigger signature verification and certificate parsing, which are not used by the application.

Updated `PackageServiceTest.kt` to include Robolectric tests that verify these flags are no longer passed.
